requirements.txt
Django
django-utils-six
django-add-default-value
django-cors-headers
django-deprecate-fields
django-easy-pdf3
django-extensions
django-field-history
django-linear-migrations
django-filter
django-migration-linter
django-timezone-field
djangorestframework
django-user-agents
django-apscheduler
djangorestframework-csv
djangorestframework-jwt
djangorestframework-yaml
drf-extensions
drf-jwt
drf-yasg
drf-typed-views
fcm-django
psycopg2-binary>=2.8
requests
beautifulsoup4